fix: harden worktree and plugin foundations
